fix code quality
